wplan/pyproject.toml
[project]
name = "wplan"
version = "0.1.0"
description = ""
authors = [
    {name = "valentin",email = "piterans.v@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "selenium (>=4.39.0,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "pyinstaller (>=6.17.0,<7.0.0)",
    "pillow (>=12.0.0,<13.0.0)"
]

[tool.poetry]
packages = [{include = "wplan", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
